+++ NEWS.md
@@ -1,6 +1,15 @@
 # rpkgkit 0.0.5 (2026-06-03)
 
+## MINOR IMPROVEMENTS
+
+* Add more options for changelog type in NEWS.md
+
+* Add more tests
+
+
 ## NEW FEATURES
+
+* Added browse_standalone() to inquire all available standalone R files across github
 
 * Added `rename_func` to change the naming style of functions